chore: 글 목록 모달 창 로직 수정
